ISIS-1976: move responsibility for pre-loading service adapters from PersistenceSession to ObjectAdapterContext
Task-Url: https://issues.apache.org/jira/browse/ISIS-1976
